Moving primary nav content into 2 components
