+++ src/components/navigation-header/components/index.ts
@@ -1,3 +1,9 @@
 import NavigationHeaderDropdownMenu from './dropdown-menu'
+import HomePageHeaderContent from './home-page-content'
+import ProductPageHeaderContent from './product-page-content'
 
-export { NavigationHeaderDropdownMenu }
+export {
+  HomePageHeaderContent,
+  NavigationHeaderDropdownMenu,
+  ProductPageHeaderContent,
+}
